Add support for deploying a bastion host
This commit adds support for deploying a bastion host in the management resource group. The `deploy_bastion` variable has been added to the network module and the necessary changes have been made in the Terraform code and pipeline templates. This allows users to choose whether to deploy a bastion host during the infrastructure provisioning process.